face-recognition-service/requirements.txt
face_recognition==1.3.0
fastapi==0.116.1
numpy==2.3.3
Pillow>=11.0.0,<12.0.0
python-dotenv==1.1.1
requests==2.32.5
uvicorn==0.35.0
